# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "投资收益统计系统 API",
    "description": "方案B — 资产快照 + 出入金 + 证券买卖 + 标的最新价 + 现金余额",
    "version": "1.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/auth/account/restore": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Restore",
        "operationId": "restore_api_auth_account_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/auth/profile": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Profile",
        "operationId": "profile_api_auth_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatchReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/auth/password": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Change Password",
        "operationId": "change_password_api_auth_password_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordPatchReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/auth/email": {
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Change Email",
        "operationId": "change_email_api_auth_email_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailPatchReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/auth/account": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Delete Account",
        "operationId": "delete_account_api_auth_account_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/portfolios/comparison": {
      "get": {
        "tags": [
          "aggregation"
        ],
        "summary": "Comparison",
        "operationId": "comparison_api_portfolios_comparison_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioSummaryOut"
                  },
                  "type": "array",
                  "title": "Response Comparison Api Portfolios Comparison Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/portfolios/{portfolio_id}/summary": {
      "get": {
        "tags": [
          "aggregation"
        ],
        "summary": "Summary",
        "operationId": "summary_api_portfolios__portfolio_id__summary_get",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/overview": {
      "get": {
        "tags": [
          "aggregation"
        ],
        "summary": "Overview",
        "operationId": "overview_api_portfolios__portfolio_id__overview_get",
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Portfolio Id"
            }
          },
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "1w|1m|3m|6m|1y|ytd|all",
              "default": "1y",
              "title": "Range"
            },
            "description": "1w|1m|3m|6m|1y|ytd|all"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/account/stats": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Account Stats",
        "operationId": "account_stats_api_account_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountStatsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/portfolios": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolios",
        "operationId": "list_portfolios_api_portfolios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PortfolioOut"
                  },
                  "type": "array",
                  "title": "Response List Portfolios Api Portfolios Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio",
        "operationId": "create_portfolio_api_portfolios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreateReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountStatsOut": {
        "properties": {
          "portfolioCount": {
            "type": "integer",
            "title": "Portfoliocount"
          },
          "totalAssets": {
            "type": "string",
            "title": "Totalassets"
          },
          "cumulativeXirr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cumulativexirr"
          },
          "yearXirr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yearxirr"
          }
        },
        "type": "object",
        "required": [
          "portfolioCount",
          "totalAssets"
        ],
        "title": "AccountStatsOut"
      },
      "AuthTokenOut": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "user": {
            "$ref": "#/components/schemas/UserPublicOut"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "user"
        ],
        "title": "AuthTokenOut"
      },
...
      "CashflowOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "type",
          "amount",
          "createdAt"
        ],
        "title": "CashflowOut"
      },
...
      "EmailPatchReq": {
        "properties": {
          "currentPassword": {
            "type": "string",
            "title": "Currentpassword"
          },
          "newEmail": {
            "type": "string",
            "title": "Newemail"
          }
        },
        "type": "object",
        "required": [
          "currentPassword",
          "newEmail"
        ],
        "title": "EmailPatchReq"
      },
      "FreshnessOut": {
        "properties": {
          "staleDays": {
            "type": "integer",
            "title": "Staledays"
          },
          "isStale": {
            "type": "boolean",
            "title": "Isstale"
          },
          "latestPriceAsOf": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestpriceasof"
          },
          "latestPriceLagDays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestpricelagdays"
          },
          "latestCashAsOf": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestcashasof"
          },
          "latestCashLagDays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestcashlagdays"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "staleDays",
          "isStale"
        ],
        "title": "FreshnessOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginReq": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginReq"
      },
      "NavPointOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "cumulativeNav": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cumulativenav"
          },
          "yearNav": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yearnav"
          },
          "shares": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shares"
          }
        },
        "type": "object",
        "required": [
          "date"
        ],
        "title": "NavPointOut",
        "description": "NAV 序列点（兼容 metric=both 的 {cumulativeNav,yearNav} 与 单值 {value}）。"
      },
      "OverviewOut": {
        "properties": {
          "totalAsset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalasset"
          },
          "cumulativeXirr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cumulativexirr"
          },
          "yearXirr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yearxirr"
          },
          "navSeries": {
            "items": {
              "$ref": "#/components/schemas/NavPointOut"
            },
            "type": "array",
            "title": "Navseries",
            "default": []
          },
          "recentCashflows": {
            "items": {
              "$ref": "#/components/schemas/CashflowOut"
            },
            "type": "array",
            "title": "Recentcashflows",
            "default": []
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessOut"
          }
        },
        "type": "object",
        "required": [
          "freshness"
        ],
        "title": "OverviewOut"
      },
...
      "PasswordPatchReq": {
        "properties": {
          "currentPassword": {
            "type": "string",
            "title": "Currentpassword"
          },
          "newPassword": {
            "type": "string",
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "title": "PasswordPatchReq"
      },
...
      "PortfolioCreateReq": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "CNY"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PortfolioCreateReq"
      },
      "PortfolioOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "baseDate": {
            "type": "string",
            "format": "date",
            "title": "Basedate"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "archivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "baseDate",
          "currency",
          "createdAt",
          "updatedAt"
        ],
        "title": "PortfolioOut"
      },
...
      "PortfolioSummaryOut": {
        "properties": {
          "cumulativeXirr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cumulativexirr"
          },
          "totalReturnRate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalreturnrate"
          },
          "yearReturnRate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yearreturnrate"
          },
          "maxDrawdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxdrawdown"
          },
          "latestDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestdate"
          },
          "inceptionDate": {
            "type": "string",
            "format": "date",
            "title": "Inceptiondate"
          }
        },
        "type": "object",
        "required": [
          "inceptionDate"
        ],
        "title": "PortfolioSummaryOut"
      },
...
      "ProfilePatchReq": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "title": "ProfilePatchReq"
      },
...
      "RegisterReq": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterReq"
      },
      "RestoreReq": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RestoreReq"
      },
...
      "UserPublicOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name"
        ],
        "title": "UserPublicOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "JWT-auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}